@@ -1,4 +1,4 @@
 export const environment = {
-  API_URL: 'https://fakestoreapi.com',
-  SERVER_URL: 'http://localhost:4242',
+  API_URL_FAKE_STORE: 'https://fakestoreapi.com',
+  SERVER_URL: 'http://localhost:4050/api',
 };
